VIC-11782 upgrade sample index counts to 64 bit (#4463)
(cherry picked from commit 1532613a5e244ba74de072e0c22459b7f7f050f7)

+++ animProcess/src/cozmoAnim/alexa/alexa.h
@@ -59,7 +59,7 @@ public:
   
   void NotifyOfTapToTalk();
   
-  void NotifyOfWakeWord( size_t fromSampleIndex, size_t toSampleIndex );
+  void NotifyOfWakeWord( uint64_t fromSampleIndex, uint64_t toSampleIndex );
   
   void SetOnCharger(bool onCharger) { _onCharger = onCharger; }
